+++ rubix-ui/frontend/src/components/connections/views/table.tsx
@@ -1,25 +1,35 @@
 import { Space, Spin } from "antd";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { PingRubixAssist } from "../../../../wailsjs/go/main/App";
 import { main, storage } from "../../../../wailsjs/go/models";
 import RbTable from "../../../common/rb-table";
-import { RbDeleteButton } from "../../../common/rb-table-actions";
+import {
+  RbAddButton,
+  RbDeleteButton,
+  RbRefreshButton,
+} from "../../../common/rb-table-actions";
 import { CONNECTION_HEADERS } from "../../../constants/headers";
 import { ROUTES } from "../../../constants/routes";
-import { openNotificationWithIcon } from "../../../utils/utils";
+import { isObjectEmpty, openNotificationWithIcon } from "../../../utils/utils";
 import { ConnectionFactory } from "../factory";
+import { CreateEditModal } from "./create";
 
 import RubixConnection = storage.RubixConnection;
 import UUIDs = main.UUIDs;
 
-export const ConnectionsTable = (props: any) => {
-  const { connections, refreshList, showModal, isFetching } = props;
-  if (!connections) return <></>;
-
-  let factory = new ConnectionFactory();
-
+export const ConnectionsTable = () => {
   const [selectedUUIDs, setSelectedUUIDs] = useState([] as Array<UUIDs>);
+  const [connections, setConnections] = useState([] as RubixConnection[]);
+  const [currentConnection, setCurrentConnection] = useState(
+    {} as RubixConnection
+  );
+  const [connectionSchema, setConnectionSchema] = useState({});
+  const [isModalVisible, setIsModalVisible] = useState(false);
+  const [isFetching, setIsFetching] = useState(true);
+  const [isLoadingForm, setIsLoadingForm] = useState(false);
+
+  const factory = new ConnectionFactory();
 
   const columns = [
     ...CONNECTION_HEADERS,
@@ -51,9 +61,49 @@ export const ConnectionsTable = (props: any) => {
     },
   ];
 
+  useEffect(() => {
+    fetch();
+  }, []);
+
+  const fetch = async () => {
+    try {
+      setIsFetching(true);
+      let res = await factory.GetAll();
+      res = !res ? [] : res;
+      setConnections(res);
+    } catch (error) {
+      setConnections([]);
+    } finally {
+      setIsFetching(false);
+    }
+  };
+
+  const getSchema = async () => {
+    setIsLoadingForm(true);
+    const res = await factory.Schema();
+    const jsonSchema = {
+      properties: res,
+    };
+    setConnectionSchema(jsonSchema);
+    setIsLoadingForm(false);
+  };
+
+  const showModal = (connection: RubixConnection) => {
+    setCurrentConnection(connection);
+    setIsModalVisible(true);
+    if (isObjectEmpty(connectionSchema)) {
+      getSchema();
+    }
+  };
+
+  const onCloseModal = () => {
+    setIsModalVisible(false);
+    setCurrentConnection({} as RubixConnection);
+  };
+
   const bulkDelete = async () => {
     await factory.BulkDelete(selectedUUIDs);
-    refreshList();
+    fetch();
   };
 
   const rowSelection = {
@@ -72,11 +122,13 @@ export const ConnectionsTable = (props: any) => {
       }
     });
 
-    refreshList();
+    fetch();
   };
 
   return (
     <div>
+      <RbRefreshButton refreshList={fetch} />
+      <RbAddButton handleClick={() => showModal({} as RubixConnection)} />
       <RbDeleteButton bulkDelete={bulkDelete} />
       <RbTable
         rowKey="uuid"
@@ -84,6 +136,15 @@ export const ConnectionsTable = (props: any) => {
         rowSelection={rowSelection}
         columns={columns}
         loading={{ indicator: <Spin />, spinning: isFetching }}
+      />
+      <CreateEditModal
+        connections={connections}
+        currentConnection={currentConnection}
+        connectionSchema={connectionSchema}
+        isModalVisible={isModalVisible}
+        isLoadingForm={isLoadingForm}
+        refreshList={fetch}
+        onCloseModal={onCloseModal}
       />
     </div>
   );
